examples/001vtts/requirements.txt
pika==1.3.2
python-dotenv==1.0.0
boto3==1.29.6
